Get blueprint working for new apps
1. Blueprint packages should have their blueprint at the "top level"
  * `index.js` should be a blueprint configuration object
  * generated files should go in root `files/` directory
2. After doing this, generation was working, but addAddonsToProject was being called with the wrong signature
