+++ files/src/resolver.js
